Fix #67 solve incorrect glusterfs backup-vol-servers
The system.glusterfs.client.volumes.salt_pki class requires
cluster_node01_address to be set to the glusterfs server address. The
prx nodes did not set this address.

Next to that the OpenStack controllers and other services set the
incorrect backup-vol-servers, because it was set to
cluster_node01_address. This patch sets the correct glusterfs
backup-vol-servers for all services.