Fix stacklightv1 with kubernetes
Currently stacklight deployment fails because mysql-wsrep package is
coming from the openstack repo. This patch fixes it by including
the openstack repo when stacklight with k8s is used on telemetry nodes.

Another issue that was solved is that system.galera.server will define
a keepalived vip that conflicts with stacklight_telemetry_vip. The fix
is to not include system.galera.server but instead utilize
system.haproxy.proxy.listen.openstack.galera directly.

Change-Id: Ifb8c46d1be028d0a58f6b1e3720903eb94445f47